refactor: create registry, cfg, and resolution packages
This PR establishes the infrastructure core by creating three new packages
that resolve circular dependencies and improve code organization. The
registry package handles module/type/attribute registries, cfg package
manages control flow graphs, and resolution package defines type resolution
structures. All changes maintain backward compatibility through type aliases.

Verification: gradle buildGo (success), gradle testGo (all pass), gradle
lintGo (0 issues), coverage 95.9-100% across new packages.

+++ sourcecode-parser/graph/callgraph/variable_extraction_test.go
@@ -398,12 +398,12 @@ def test():
 	xBinding := scope.Variables["x"]
 	assert.NotNil(t, xBinding)
 	assert.Equal(t, filePath, xBinding.Location.File)
-	assert.Equal(t, 3, xBinding.Location.Line)
+	assert.Equal(t, uint32(3), xBinding.Location.Line)
 
 	yBinding := scope.Variables["y"]
 	assert.NotNil(t, yBinding)
 	assert.Equal(t, filePath, yBinding.Location.File)
-	assert.Equal(t, 4, yBinding.Location.Line)
+	assert.Equal(t, uint32(4), yBinding.Location.Line)
 }
 
 // TestInferTypeFromExpression_DirectCalls tests type inference helper.
